Move `state/game_state.rs` to `state/game.rs`
The original file name was redundant.

I also made the `game`/`game_state` module private again. The reason
making it public fixed it before is because you were accessing game
through its actual module in `main.rs` instead of accessing it through
the re-export in the state module. (See the diff on `main.rs` if that's
confusing.)